+++ CHANGELOG.rst
@@ -3,6 +3,18 @@ Changelog
 ==========
 
 ++++++++++
+v0.3.0 (23/01/2023)
+++++++++++
+
+**Added**
+
+- Data collection statistics page for single particle analysis
+
+**Changed**
+
+- Every request displays loading bar, instead of loading bar state being changed by hand
+
+++++++++++
 v0.2.0 (23/01/2023)
 ++++++++++